Upgraded NotebookService to Angular. #2479
