+++ src/main/webapp/wise5/classroomMonitor/classroomMonitorComponents/notebook/notebookWorkgroupGrading/notebookWorkgroupGrading.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import NotebookService from '../../../../services/notebookService';
+import { NotebookService } from '../../../../services/notebookService';
 import { TeacherProjectService } from '../../../../services/teacherProjectService';
 
 class NotebookWorkgroupGradingController {
