+++ src/main/webapp/wise5/classroomMonitor/classroomMonitorController.ts
@@ -2,7 +2,7 @@
 
 import { ConfigService } from '../services/configService';
 import NodeService from '../services/nodeService';
-import NotebookService from '../services/notebookService';
+import { NotebookService } from '../services/notebookService';
 import NotificationService from '../services/notificationService';
 import TeacherDataService from '../services/teacherDataService';
 import { SessionService } from '../services/sessionService';
