+++ src/main/webapp/wise5/classroomMonitor/notebook/notebookGradingController.ts
@@ -2,7 +2,7 @@
 
 import * as angular from 'angular';
 import { ConfigService } from '../../services/configService';
-import NotebookService from '../../services/notebookService';
+import { NotebookService } from '../../services/notebookService';
 import TeacherDataService from '../../services/teacherDataService';
 
 class NotebookGradingController {
