+++ src/main/webapp/wise5/components/componentController.ts
@@ -3,7 +3,7 @@ import * as $ from 'jquery';
 import { AnnotationService } from "../services/annotationService";
 import { ConfigService } from "../services/configService";
 import NodeService from "../services/nodeService";
-import NotebookService from "../services/notebookService";
+import { NotebookService } from "../services/notebookService";
 import { ProjectService } from "../services/projectService";
 import { StudentAssetService } from "../services/studentAssetService";
 import { UtilService } from "../services/utilService";
