+++ src/main/webapp/wise5/vle/vleController.ts
@@ -3,7 +3,7 @@
 import { AnnotationService } from '../services/annotationService';
 import { ConfigService } from '../services/configService';
 import NotificationService from '../services/notificationService';
-import NotebookService from '../services/notebookService';
+import { NotebookService } from '../services/notebookService';
 import { VLEProjectService } from './vleProjectService';
 import { SessionService } from '../services/sessionService';
 import { StudentDataService } from '../services/studentDataService';
